removes the old deployment test from test_ocs_basic_install.py
This also ports all changes from that back into the cluster fixture
